+++ packages/recorder/src/index.ts
@@ -1,8 +1,15 @@
 import {initApp} from './express_app';
-import {SessionManager} from '../session_recording.ts/sessions_manager';
+import {SessionManager} from './session_recording.ts/sessions_manager';
+
+const defaultDestination = process.env.DESTINATION;
+const useDestinationHostHeader = process.env.USE_DESTINATION_HOST_HEADER;
+
+if (!defaultDestination && useDestinationHostHeader !== 'true') {
+    console.log('Please provide a destination via the DESTINATION enviornment variable, or bypass this check by setting the USE_DESTINATION_HOST_HEADER bool environment variable');
+}
 
 const sessionManager = new SessionManager();
-const app = initApp({sessionManager, shouldProxy: true});
+const app = initApp({sessionManager, defaultDestination});
 
 const port = process.env.PORT || 3000;
