+++ packages/recorder/src/session_recording.ts/session_recorder.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs/promises';
 
-import {RecordedCall} from '../src/types';
+import {RecordedCall} from '../types';
 
 export class RecordingSession {
     private startTime: Date = new Date();
