+++ packages/recorder/src/session_recording.ts/sessions_manager.ts
